tfpose-realtime2/requirements2.txt
Pillow
cython
tf_pose
scikit-image
opencv-python
h5py
imgaug
pandas
argparse
IPython[all]
simplejson
sklearn
slidingwindow
pyyaml
matplotlib
